Use `serde_dynamo` for `dynamodb::Event` items
Use `serde_dynamo::Item` instead of `HashMap<String, AttributeValue>` to
represent the `keys`, `old_image`, and `new_image` in a
`dynamodb::Event`.

This consolidates the Rust/DynamoDB ecosystem around fewer
representations of DynamoDB items.

Note that `serde_dynamo::AttributeValue` and
`aws_lambda_events::dynamodb::attributes::AttributeValue` are slightly
different (especially with regards to the `Number` enumeration) so this
is a breaking change that would either need to go behind a feature flag
or a major version bump.